+++ src/main/java/com/example/Feline.java
@@ -2,7 +2,7 @@ package com.example;
 
 import java.util.List;
 
-public class Feline extends Animal implements Predator, Kittens {
+public class Feline extends Animal implements Predator {
 
     @Override
     public List<String> eatMeat() throws Exception {
@@ -14,12 +14,10 @@ public class Feline extends Animal implements Predator, Kittens {
         return "Кошачьи";
     }
 
-    @Override
     public int getKittens() {
         return getKittens(1);
     }
 
-    @Override
     public int getKittens(int kittensCount) {
         return kittensCount;
     }
