--- src/main/java/com/example/Kittens.java
@@ -1,9 +1,0 @@
-package com.example;
-
-public interface Kittens {
-
-    int getKittens();
-
-    int getKittens(int kittensCount);
-
-}
