+++ src/main/java/com/example/Predator.java
@@ -6,4 +6,6 @@ public interface Predator {
 
     List<String> eatMeat() throws Exception;
 
+    int getKittens();
+
 }
