+++ src/test/java/com/example/CatTest.java
@@ -1,22 +1,30 @@
 package com.example;
 
 import org.junit.Test;
-
 import org.junit.runner.RunWith;
 import org.mockito.Mock;
 import org.mockito.Mockito;
 import org.mockito.junit.MockitoJUnitRunner;
 
-
-import java.util.Collections;
+import java.util.ArrayList;
+import java.util.Arrays;
 
 import static org.junit.Assert.assertEquals;
 
 @RunWith(MockitoJUnitRunner.class)
-public class CatTestGetSoundValue {
+public class CatTest {
 
     @Mock
     Feline feline;
+
+    @Test
+    public void CatTestGetFootValidValue() throws Exception {
+        ArrayList<String> result = new ArrayList<>(Arrays.asList("Животные", "Птицы", "Рыба"));
+
+        Cat cat = new Cat(feline);
+        Mockito.when(feline.eatMeat()).thenReturn(result);
+        assertEquals(result, cat.getFood());
+    }
 
     @Test
     public void CatTestGetSoundValidValue() {
@@ -24,6 +32,7 @@ public class CatTestGetSoundValue {
         assertEquals("Мяу", cat.getSound());
 
     }
+
 
 }
