--- src/test/java/com/example/FelineTestGetFamilyValue.java
@@ -1,19 +1,0 @@
-package com.example;
-
-import org.junit.Test;
-
-import static org.junit.Assert.assertEquals;
-
-
-public class FelineTestGetFamilyValue {
-
-    @Test
-    public void FelineTestGetFamilyValidValue() {
-        Feline feline = new Feline();
-        assertEquals("Кошачьи", feline.getFamily());
-
-    }
-
-}
-
-
