+++ src/test/java/com/example/LionTestSexException.java
@@ -1,5 +1,4 @@
 package com.example;
-import com.example.Lion;
 import org.junit.Assert;
 import org.junit.Test;
 import org.junit.runner.RunWith;
