+++ src/test/java/com/example/LionTestSexValue.java
@@ -1,5 +1,4 @@
 package com.example;
-import com.example.Lion;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.junit.runners.Parameterized;
